% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenotypeEffect.R
\name{plotGenotypeEffect}
\alias{plotGenotypeEffect}
\title{Plot phenotype values grouped by SNP genotype}
\usage{
plotGenotypeEffect(
  x,
  snp_id,
  phenotype_id,
  assayName = NULL,
  size = 2,
  color = "steelblue",
  title = NULL
)
}
\arguments{
\item{x}{A \link{tQTLExperiment}.}

\item{snp_id}{Character. SNP identifier (row name in \code{variantRanges(x)}).}

\item{phenotype_id}{Character. Phenotype/gene identifier (row name in
\code{rowRanges(x)}).}

\item{assayName}{Name of the assay to use. Defaults to the first assay.}

\item{size}{Point size for beeswarm. Default 2.}

\item{color}{Color for points. Default "steelblue".}

\item{title}{Main plot title. If \code{NULL}, auto-generated from SNP and gene.}
}
\value{
A ggplot2 plot object.
}
\description{
Creates a beeswarm plot comparing a phenotype across genotype groups
(0, 1, 2 copies of the alternate allele) for a given SNP.  Useful for
visualizing the effect of a significant eQTL variant.
}
\examples{
data(mageSEfilt, package="CSHLvc2026")
plink_paths <- cache_mage_chr17_plink()
plpre <- tools::file_path_sans_ext(plink_paths[["bed"]])
cd <- as.data.frame(colData(mageSEfilt))
mm <- model.matrix(~ batch + population + sex, data = cd)
mm <- mm[, -1, drop = FALSE]   # remove (Intercept)
tqe <- tQTLExperimentFromRSE(
    se              = mageSEfilt,
    plinkPrefix     = plpre,
    covariateMatrix = mm,
    genome          = "hg38"
)
plotGenotypeEffect(tqe, "17:410391:T:C", "ENSG00000187624.9")
}
