% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,tQTLExperiment-method}
\alias{show,tQTLExperiment-method}
\title{Show a tQTLExperiment}
\usage{
\S4method{show}{tQTLExperiment}(object)
}
\arguments{
\item{object}{A \link{tQTLExperiment}.}
}
\description{
Show a tQTLExperiment
}
