% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[,tQTLExperiment,ANY,ANY,ANY-method}
\alias{[,tQTLExperiment,ANY,ANY,ANY-method}
\title{Subset a tQTLExperiment}
\usage{
\S4method{[}{tQTLExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \link{tQTLExperiment}.}

\item{i}{Row (feature) index.}

\item{j}{Column (sample) index.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\value{
A \link{tQTLExperiment}.
}
\description{
Sample subsetting propagates to the lazy \code{geno} matrix (BEDMatrix rows).
}
\examples{
exdir <- system.file("extdata", package = "tQTLExperiment")
tqe <- tQTLExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed")
)
tqe[1:5, 1:10]
}
