% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTQTL.R
\name{tensorqtlVersion}
\alias{tensorqtlVersion}
\title{Report the tensorQTL version from a Python installation}
\usage{
tensorqtlVersion(python = findTQTL())
}
\arguments{
\item{python}{Path to the Python executable. Defaults to \code{\link[=findTQTL]{findTQTL()}}.}
}
\value{
A character string with the version, or \code{NA}.
}
\description{
Report the tensorQTL version from a Python installation
}
\examples{
tryCatch(tensorqtlVersion(), error = function(e) NA_character_)
}
