% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{tqtlGeno}
\alias{tqtlGeno}
\alias{tqtlGeno<-}
\alias{tqtlGeno,tQTLExperiment-method}
\alias{tqtlGeno<-,tQTLExperiment-method}
\title{Access the lazy genotype matrix}
\usage{
tqtlGeno(x)

tqtlGeno(x) <- value

\S4method{tqtlGeno}{tQTLExperiment}(x)

\S4method{tqtlGeno}{tQTLExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{tQTLExperiment}.}

\item{value}{A \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix} (samples × variants).}
}
\value{
The \code{geno} slot (samples × variants BEDMatrix).
}
\description{
Access the lazy genotype matrix
}
\examples{
exdir <- system.file("extdata", package = "tQTLExperiment")
tqe <- tQTLExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed")
)
tqtlGeno(tqe)
}
