% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{tqtlPlinkPrefix}
\alias{tqtlPlinkPrefix}
\alias{tqtlPlinkPrefix<-}
\alias{tqtlPlinkPrefix,tQTLExperiment-method}
\alias{tqtlPlinkPrefix<-,tQTLExperiment-method}
\title{Access the PLINK file prefix}
\usage{
tqtlPlinkPrefix(x)

tqtlPlinkPrefix(x) <- value

\S4method{tqtlPlinkPrefix}{tQTLExperiment}(x)

\S4method{tqtlPlinkPrefix}{tQTLExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{tQTLExperiment}.}

\item{value}{A length-one character string.}
}
\value{
The PLINK prefix path.
}
\description{
Access the PLINK file prefix
}
\examples{
exdir <- system.file("extdata", package = "tQTLExperiment")
tqe <- tQTLExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed")
)
tqtlPlinkPrefix(tqe)
}
