% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{tqtlVariantRanges}
\alias{tqtlVariantRanges}
\alias{tqtlVariantRanges<-}
\alias{tqtlVariantRanges,tQTLExperiment-method}
\alias{tqtlVariantRanges<-,tQTLExperiment-method}
\title{Access variant genomic ranges}
\usage{
tqtlVariantRanges(x)

tqtlVariantRanges(x) <- value

\S4method{tqtlVariantRanges}{tQTLExperiment}(x)

\S4method{tqtlVariantRanges}{tQTLExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{tQTLExperiment}.}

\item{value}{A \link[GenomicRanges:GRanges]{GenomicRanges::GRanges}.}
}
\value{
A \code{GRanges} with metadata columns \code{snp_id}, \code{ref}, \code{alt}.
}
\description{
Access variant genomic ranges
}
\examples{
exdir <- system.file("extdata", package = "tQTLExperiment")
tqe <- tQTLExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed")
)
tqtlVariantRanges(tqe)
}
