% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTQTL.R
\name{tensorqtlVersion}
\alias{tensorqtlVersion}
\title{Report the installed tensorQTL version}
\usage{
tensorqtlVersion(python = findTQTL())
}
\arguments{
\item{python}{Path to the Python executable. Defaults to the result of
\code{\link[=findTQTL]{findTQTL()}}.}
}
\value{
A character string with the tensorQTL version, or \code{NA} if not
found.
}
\description{
Report the installed tensorQTL version
}
\examples{
tryCatch(tensorqtlVersion(), error = function(e) NA_character_)
}
